Add [FormHandlerUtil]: create a new class for storing public and utils methods. Note: Perhaps object fit better for this role
